refactor(card-browser): move showChangeDeckDialog to Fragment
